Cleaned up make_fisher_clust and added a utils file
